7. Generate controller framework logic with macro
Similar to the endpoint, the implementation details for the controller
pipeline can be generated with a macro. But be warned, there's a
problem! Plugs are run in the order they're defined and this
implementation runs `apply_action` before anything else... See the next
commit.
