+++ lib/your_app/controller.ex
@@ -1,16 +1,5 @@
 defmodule YourApp.Controller do
-  use Plug.Builder
-  plug(:apply_action)
-
-  def call(conn, action) do
-    conn
-    |> put_private(:action, action)
-    |> super(nil)
-  end
-
-  def apply_action(conn, _opts) do
-    apply(__MODULE__, conn.private.action, [conn])
-  end
+  use Feenix.Controller
 
   def index(conn) do
     send_resp(conn, 200, "meows")
